requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-python-client==1.7.9
google-auth==2.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.20.4
idna==3.4
mdutils==1.4.0
oauthlib==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.12
